adjusted costs for increased daylength

+++ src/base_station_base.h
@@ -54,7 +54,7 @@ struct StationRect : public Rect {
 struct BaseStation : StationPool::PoolItem<&_station_pool> {
 	TileIndex xy;                   ///< Base tile of the station
 	ViewportSign sign;              ///< NOSAVE: Dimensions of sign
-	byte delete_ctr;                ///< Delete counter. If greater than 0 then it is decremented until it reaches 0; the waypoint is then is deleted.
+	uint16 delete_ctr;                ///< Delete counter. If greater than 0 then it is decremented until it reaches 0; the waypoint is then is deleted.
 
 	char *name;                     ///< Custom name
 	StringID string_id;             ///< Default name (town area) of station
